Move proxy-func to utils
