+++ src/main/assisted-installer-controller/assisted_installer_main.go
@@ -3,14 +3,11 @@ package main
 import (
 	"context"
 	"log"
-	"net/http"
-	"net/url"
-	"os"
 	"sync"
 	"time"
 
 	"github.com/openshift/assisted-installer/src/k8s_client"
-	"golang.org/x/net/http/httpproxy"
+	"github.com/openshift/assisted-installer/src/utils"
 
 	"github.com/kelseyhightower/envconfig"
 	assistedinstallercontroller "github.com/openshift/assisted-installer/src/assisted_installer_controller"
@@ -22,11 +19,6 @@ import (
 var Options struct {
 	ControllerConfig assistedinstallercontroller.ControllerConfig
 }
-
-var (
-	envProxyOnce          sync.Once
-	envVarsProxyFuncValue func(*url.URL) (*url.URL, error)
-)
 
 func main() {
 	logger := logrus.New()
@@ -51,7 +43,7 @@ func main() {
 
 	client, err := inventory_client.CreateInventoryClient(Options.ControllerConfig.ClusterID,
 		Options.ControllerConfig.URL, Options.ControllerConfig.PullSecretToken, Options.ControllerConfig.SkipCertVerification,
-		Options.ControllerConfig.CACertPath, logger, ProxyFromEnvVars)
+		Options.ControllerConfig.CACertPath, logger, utils.ProxyFromEnvVars)
 	if err != nil {
 		log.Fatalf("Failed to create inventory client %v", err)
 	}
@@ -87,27 +79,4 @@ func main() {
 	logger.Infof("All routines finished, closing logs")
 	cancelLogs()
 	wgLogs.Wait()
-}
-
-// ProxyFromEnvVars provides an alternative to http.ProxyFromEnvironment since it is being initialized only
-// once and that happens by k8s before proxy settings was obtained. While this is no issue for k8s, it prevents
-// any out-of-cluster traffic from using the proxy
-func ProxyFromEnvVars(req *http.Request) (*url.URL, error) {
-	return envVarsProxyFunc()(req.URL)
-}
-
-func envVarsProxyFunc() func(*url.URL) (*url.URL, error) {
-	envProxyOnce.Do(func() {
-		envVarsProxyFuncValue = FromEnvVars().ProxyFunc()
-	})
-	return envVarsProxyFuncValue
-}
-
-func FromEnvVars() *httpproxy.Config {
-	return &httpproxy.Config{
-		HTTPProxy:  os.Getenv("HTTP_PROXY"),
-		HTTPSProxy: os.Getenv("HTTPS_PROXY"),
-		NoProxy:    os.Getenv("NO_PROXY"),
-		CGI:        os.Getenv("REQUEST_METHOD") != "",
-	}
 }
